IGNITE-24749 review fixes
