+++ modules/distribution-zones/build.gradle
@@ -102,7 +102,6 @@ dependencies {
     integrationTestImplementation project(':ignite-system-view-api')
     integrationTestImplementation project(':ignite-failure-handler')
     integrationTestImplementation project(':ignite-system-disaster-recovery')
-    integrationTestImplementation project(':ignite-configuration-system')
     integrationTestImplementation project(':ignite-transactions')
     integrationTestImplementation project(':ignite-raft-api')
     integrationTestImplementation project(':ignite-raft')
